Move OutputFormat to own module
